Refactor InfinispanCacheConfiguration from wildfly-clustering-web-infinispan to wildfly-clustering-ee-infinispan.
