+++ clustering/ee/infinispan/src/main/java/org/wildfly/clustering/ee/infinispan/InfinispanCacheConfiguration.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.web.infinispan;
+package org.wildfly.clustering.ee.infinispan;
 
 /**
  * General configuration that identifies an Infinispan cache.
