+++ clustering/web/extension/src/main/java/org/wildfly/extension/clustering/web/routing/infinispan/InfinispanRoutingConfiguration.java
@@ -25,7 +25,7 @@ package org.wildfly.extension.clustering.web.routing.infinispan;
 import java.util.function.Consumer;
 
 import org.infinispan.configuration.cache.ConfigurationBuilder;
-import org.wildfly.clustering.web.infinispan.InfinispanCacheConfiguration;
+import org.wildfly.clustering.ee.infinispan.InfinispanCacheConfiguration;
 
 /**
  * Configuration of an Infinispan routing provider
