+++ clustering/web/extension/src/main/java/org/wildfly/extension/clustering/web/routing/infinispan/PrimaryOwnerRouteLocatorServiceConfiguratorFactory.java
@@ -23,8 +23,8 @@
 package org.wildfly.extension.clustering.web.routing.infinispan;
 
 import org.jboss.as.clustering.controller.CapabilityServiceConfigurator;
+import org.wildfly.clustering.ee.infinispan.InfinispanCacheConfiguration;
 import org.wildfly.clustering.web.WebDeploymentConfiguration;
-import org.wildfly.clustering.web.infinispan.InfinispanCacheConfiguration;
 import org.wildfly.extension.clustering.web.routing.RouteLocatorServiceConfiguratorFactory;
 
 /**
